+++ androiddi/src/main/java/com/mission/androiddi/component/activity/retain/RetainedActivityDependencyLifecycleObserver.kt
@@ -1,11 +1,12 @@
-package com.mission.androiddi.component.activity
+package com.mission.androiddi.component.activity.retain
 
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.LifecycleOwner
+import com.mission.androiddi.component.activity.InjectableActivity
 import com.mission.androiddi.component.application.InjectableApplication
 import com.woowacourse.bunadi.injector.Injector
 
-class ActivityDependencyLifecycleObserver(
+class RetainedActivityDependencyLifecycleObserver(
     private val injector: Injector,
     private val activity: InjectableActivity,
 ) : DefaultLifecycleObserver {
